Script to extract co-authorship matrix
So far extracted binary matricies (e.g., 1=published together in that year) for cogsci, nips and vss, there is a function in the .R file for extracting the counts, will run later. You need to load in the _byAuthor.csv for all three data sets for the script to work.